Desire.Core - Modificador.cs: Classe extensivamente alterada para ser facilmente expandível (SOLID). - Adicionados os tipos de modificadores: ModHabilidade, ModValorMag, ModNumero, ModPericia - CalculadorNumero.cs: adicionada função "SubtraiValorMag" - Outras classes e métodos alterados para refletir alterações na classe de modificadores. - Variáveis do tipo "double" mudadas para "decimal" para facilitar a modificação.
Desire.Test
- Adicionados testes para alterações de classe de modificadores

+++ Desire.Core/Energias/Energia.cs
@@ -3,9 +3,9 @@
     public class Energia
     {
         public string Sigla { get; set; }
-        public double Quantidade { get; set; }
+        public decimal Quantidade { get; set; }
 
-        public Energia(string sigla, double quantidade)
+        public Energia(string sigla, decimal quantidade)
         {
             this.Sigla = sigla;
             this.Quantidade = quantidade;
